backend/requirements.txt
fastapi[all]
uvicorn
python-dotenv
aiohttp
supabase-client
sqlalchemy
psycopg2-binary
google-generativeai
alembic